Manuelle Korrektur implementiert
- Manuelle Korrektur für automatisch markierte Aufgaben implementiert
- ManuelleKorrekturGUI erstellt
- ManuelleKorrekturService hinzugefügt
- Bewertungsengine auf AufgabeErgebnis umgestellt
- Aufgabenstatus für automatische und manuelle Bewertung eingeführt
- Neuberechnung von Gesamtpunkten und Prozentwert nach manueller Korrektur
- Integration der manuellen Korrektur in den Dozentenbereich

+++ code/model/AufgabeErgebnis.java
@@ -43,7 +43,7 @@ public class AufgabeErgebnis {
         return punkte;
     }
 
-    public void setPunkte(int punkte) {
+    public void setPunkte(double punkte) {
         this.punkte = punkte;
     }
